+++ frontend/package.json
@@ -24,6 +24,7 @@
     "papaparse": "^5.3.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-player": "^2.11.0",
     "siwe": "^1.1.6",
     "wagmi": "^0.6.7"
   },
