+++ frontend/src/components/SideBar.tsx
@@ -1,6 +1,12 @@
 import { Dialog, Transition } from '@headlessui/react'
 import { Fragment } from 'react'
-import { HandRaisedIcon, HomeIcon, InformationCircleIcon, XMarkIcon } from '@heroicons/react/24/outline'
+import {
+  HandRaisedIcon,
+  HomeIcon,
+  InformationCircleIcon,
+  VideoCameraIcon,
+  XMarkIcon
+} from '@heroicons/react/24/outline'
 import { classNames } from '../utils'
 import { ConnectButton } from './ConnectButton'
 import { useRouter } from 'next/router'
@@ -16,6 +22,12 @@ const SideBar = ({
       name: 'News',
       path: '/',
       icon: HomeIcon,
+      className: 'hover:cursor-pointer'
+    },
+    {
+      name: 'Videos',
+      path: 'videos',
+      icon: VideoCameraIcon,
       className: 'hover:cursor-pointer'
     },
     {
